cleanup public scala api
