+++ src/main/scala/akka/persistence/dynamodb/query/javadsl/DynamodbReadJournal.scala
@@ -1,7 +1,6 @@
 package akka.persistence.dynamodb.query.javadsl
 
 import akka.NotUsed
-import akka.persistence.dynamodb.query.scaladsl.{ DynamodbReadJournal => ScalaDynamodbReadJournal }
 import akka.persistence.query.EventEnvelope
 import akka.stream.javadsl.Source
 
@@ -26,13 +25,13 @@ object DynamodbReadJournal {
  *   PersistenceQuery.get(system).getReadJournalFor(DynamodbReadJournal.class, DynamodbReadJournal.Identifier());
  * }}}
  *
- * Corresponding Scala API is in [[akka.persistence.dynamodb.query.scaladsl.DynamodbReadJournal]].
+ * Corresponding Scala API is in [[DynamodbReadJournal]].
  *
  * Configuration settings can be defined in the configuration section with the
  * absolute path corresponding to the identifier, which is `"dynamodb-read-journal"`
  * for the default [[DynamodbReadJournal#Identifier]]. See `reference.conf`.
  */
-class DynamodbReadJournal(scaladslReadJournal: ScalaDynamodbReadJournal)
+class DynamodbReadJournal(scaladslReadJournal: DynamodbReadJournal)
     extends akka.persistence.query.javadsl.ReadJournal
     //    with akka.persistence.query.javadsl.EventsByTagQuery
     with akka.persistence.query.javadsl.CurrentEventsByPersistenceIdQuery
