+++ src/main/scala/akka/persistence/dynamodb/query/scaladsl/DynamodbReadJournal.scala
@@ -1,15 +1,5 @@
 package akka.persistence.dynamodb.query.scaladsl
 
-import akka.actor.ExtendedActorSystem
-import akka.actor.TypedActor.context
-import akka.persistence.dynamodb.journal._
-import akka.persistence.dynamodb.query.{ DynamoDBReadJournalConfig, ReadJournalSettingsProvider }
-import akka.persistence.dynamodb._
-import akka.persistence.query.scaladsl.ReadJournal
-import akka.serialization.{ Serialization, SerializationExtension }
-import akka.stream.{ Materializer, SystemMaterializer }
-import com.typesafe.config.Config
-
 /**
  * Scala API `akka.persistence.query.scaladsl.ReadJournal` implementation for Dynamodb.
  *
@@ -24,27 +14,7 @@ import com.typesafe.config.Config
  * absolute path corresponding to the identifier, which is `"dynamodb-read-journal"`
  * for the default [[DynamodbReadJournal#Identifier]]. See `reference.conf`.
  */
-class DynamodbReadJournal(config: Config, configPath: String)(implicit val system: ExtendedActorSystem)
-    extends ReadJournal
-    with DynamodbCurrentPersistenceIdsQuery
-    with DynamodbCurrentEventsByPersistenceIdQuery
-    with ReadJournalSettingsProvider
-    with JournalSettingsProvider
-    with DynamoProvider
-    with ActorSystemProvider
-    with MaterializerProvider
-    with LoggingProvider
-    with JournalKeys
-    with SerializationProvider
-    with ActorSystemLoggingProvider {
-
-  protected val readJournalSettings       = new DynamoDBReadJournalConfig(config)
-  protected val dynamo: DynamoDBHelper    = dynamoClient(system, readJournalSettings)
-  val serialization: Serialization        = SerializationExtension(system)
-  implicit val materializer: Materializer = SystemMaterializer(context.system).materializer
-  val journalSettings                     = new DynamoDBJournalConfig(config)
-
-}
+trait DynamodbReadJournal extends DynamodbCurrentEventsByPersistenceIdQuery with DynamodbCurrentPersistenceIdsQuery
 
 object DynamodbReadJournal {
